Update LinearResponse.m
Added first order implementation of the LinearResponse function.

The output is somewhat overloaded now. In the first order case the output is still called x in the function, but now refers to the vector z in the paper containing positions and velocities. To make this clear, the quantity z is explicitly assigned to x before outputting.