Implementation in first order (#4)
* Update SSR.m

Fixed the bug reported earlier.

The errors produced in the modified shaw pierre example prior to this bugfix do not occur anymore. Additionally the results match with the first order implementation

* Updated Iteration functions

Added the first order implementation in the functions used for the iterations.

* Create G.m

Added the Greens function for the first order Implementation

* Update convolution_x.m

Added the first order implementation for the convolution.

* Update LinearResponse.m

Added first order implementation of the LinearResponse function.

The output is somewhat overloaded now. In the first order case the output is still called x in the function, but now refers to the vector z in the paper containing positions and velocities. To make this clear, the quantity z is explicitly assigned to x before outputting.

* Update NewtonRaphson.m

Added first order implementation.

* Added/Updated not_updated functions

Added the first order implementation in the functions that make sure quantities get updated

* Update Picard.m

Added first order implementation to the picard iteration function.

* Create update_Gvec.m

Added function that updates the Greens-function in the first order implementation

* Create update_Hmat.m

Added function of amplification factors for the first order implementation.

* Updated the update_.. functions

Added the first order implementation to the update functions.

* Update SSR.m - added properties

Added properties necessary for first order implementation.
Changed the name of the modified convolution matrix from A to ConvMtx_A !

* Update SSR.m - SSR function first order

Added first order implementation to the function that sets up the system. Now this function takes an inputstruct telling it the order of the system and giving the system matrices.

* Update SSR.m - added methods

Added the methods used in the first order implementation

* SSR.m - bugfix

* Bugfix in the function and its derivative used for iteration

Corrected minor bugs

* Bugfix in Greens function and Linear Response

* Update NewtonRaphson.m - Bugfix

Fixed small bug to avoid issues when iteration does not converge

* Added first order examples

added the examples that I implemented in first order form to confirm behaviour

* Updated Examples Input

Updated the way the system is input to the SSR package in the existent examples

* Update modeselect.m

Updated the modeselection function to give right input to the SSR package

* Update SSR.m - bugfix

update of small bug where system quantities were not referenced correctly

* amending the Vinv (now W) property for the general case of non symmetric input

* minor fixes in first-order examples

Co-authored-by: thuthoma <thuthoma@student.ethz.ch>